requirements.txt
selenium==4.1.3
python-dotenv~=0.20.0